SGBoxMargin is now CGFloat

+++ fe-aqua/SGBoxView.h
@@ -46,7 +46,7 @@ typedef enum
     SGBoxOrderLIFO,
 }	SGBoxOrder;
 
-typedef short SGBoxMargin;
+typedef CGFloat SGBoxMargin;
 
 @interface SGBoxView : SGView
 {
